+++ ts/packages/providers/langchain/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @composio/langchain
 
+## 0.1.55
+
+### Patch Changes
+
+- Updated dependencies [8741165]
+  - @composio/core@0.1.55
+
 ## 0.1.54
 
 ### Patch Changes