+++ ts/packages/providers/vercel/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @composio/vercel
 
+## 0.2.11
+
+### Patch Changes
+
+- Updated dependencies [8741165]
+  - @composio/core@0.1.55
+
 ## 0.2.10
 
 ### Patch Changes